Added more precise increment selection - Any double rounded up to the tenth is possible - Copied the commit logic from other input fields

+++ lib/components/Global/Phase.dart
@@ -18,7 +18,7 @@ class TrainingStage {
   int reps;
 
   @HiveField(3)
-  int increment;
+  double increment;
   
   @HiveField(4)
   List<BreathingPhase> breathingPhases;
